Added ignore option for merge requests
Also added username to the email that is sent, as well as clarified the
action that is taken by each merge request command.

+++ controllers/merge_requests_controller.php
@@ -100,6 +100,7 @@ class MergeRequestsController extends AppController {
 			$this->MergeRequest->delete($id);
 			$this->Session->setFlash('Merge was successful.', 'flash'.DS.'success');
 			
+			$this->set('user', $Model->read(null, $request['MergeRequest']['merge_id']));
 			$this->Notifier->notify(array(
 				'to' => $request['MergeRequest']['merge_id'],
 				'subject' => 'Your account has been activated',
@@ -119,8 +120,9 @@ class MergeRequestsController extends AppController {
  * Delete a request
  *
  * @param integer $id The id of the request to delete
+ * @param boolean $ignore Whether or not to just ignore it
  */
-	function delete($id = null) {		
+	function delete($id = null, $ignore = 0) {		
 		if (!$id) {
 			$this->cakeError('error404');
 		}
@@ -130,13 +132,31 @@ class MergeRequestsController extends AppController {
 		// get model we're merging
 		$Model = ClassRegistry::init($request['MergeRequest']['model']);
 		
-		// delete associated model target first
-		if ($Model->delete($request['MergeRequest']['model_id'])) {
-			// remove request
+		if ($ignore) {
+			// activate new user
+			$Model->id = $request['MergeRequest']['model_id'];
+			if ($Model->hasField('active')) {
+				$Model->saveField('active', true);
+			}
+			
+			$this->set('user', $Model->read());
+			$this->Notifier->notify(array(
+				'to' => $request['MergeRequest']['model_id'],
+				'subject' => 'Your account has been activated',
+				'template' => 'merge_requests_merge'
+			), 'email');
+			
 			$this->MergeRequest->delete($id);
-			$this->Session->setFlash('Merge request was deleted.', 'flash'.DS.'success');
-			$this->redirect(array('action'=>'index', 'model' => $request['MergeRequest']['model']));
+		} else {
+			// delete associated model target first
+			if ($Model->delete($request['MergeRequest']['model_id'])) {
+				// remove request
+				$this->MergeRequest->delete($id);
+				$this->Session->setFlash('Merge request was deleted.', 'flash'.DS.'success');
+				$this->redirect(array('action'=>'index', 'model' => $request['MergeRequest']['model']));
+			}
 		}
+		
 		$this->Session->setFlash('Unable to process request. Please try again.', 'flash'.DS.'failure');
 		$this->redirect(array('action' => 'index', 'model' => $request['MergeRequest']['model']));
 	}
